requirements.txt
streamlit==1.32.0
google-cloud-bigquery==3.17.2
pandas-gbq==0.21.0
requests==2.31.0
pandas==2.2.0
python-dotenv==1.0.1
